backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Team Management Backend API"
authors = ["Haig Santourian <haig.santourian@gmail.com>"]
readme = "README.md"

# This explicitly tells Poetry this is not a package
packages = []

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = ">=5.1.7,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

